# fix!: migrate `argo_archived_workflows.workflow` to `jsonb`