@agilgur5 2024-10-17
I've marked the title as breaking here (`fix!`) as it is a schema change and requires a long migration, [as you mentioned](https://github.com/argoproj/argo-workflows/issues/13601#issuecomment-2420499551)

> Fixes #13601

I also don't think this fully fixes #13601 since it does not affect MySQL, is an order of magnitude gain over JSON in Postgres, but can still take multiple seconds, and doesn't resolve the description of the issue either, but another related performance issue brought up in the thread