@MasonM 2024-10-18
Also
> but [can still take multiple seconds](https://github.com/argoproj/argo-workflows/issues/13601#issuecomment-2376579589)

I'm pretty sure that's solely due to the JSON cast issue fixed in https://github.com/argoproj/argo-workflows/pull/13777. I spent over a week trying to reproduce that and I can't.