@agilgur5 2024-10-18
> I did try benchmarking the index optimization, but the performance improvement was very small, which is why I didn't include it here:

It would still be better and less complex than a whole subquery IMO

> I'm pretty sure that's solely due to the JSON cast issue fixed in #13777 (and this PR)

You believe the cast itself adding multiple seconds to a query? Certainly possible, but that would be even more surprising than the index misses or existing JSON inefficiencies with this issue in the past 😅