@agilgur5 2024-10-18
> It slows it down by over 2x on my machine.

The cast alone? 👀 👀 As far as I could tell, here and in https://github.com/argoproj/argo-workflows/pull/13777 you mentioned both removing the cast _and_ moving to JSONB

> But you don't have to take my word for it, because you can reproduce this yourself by following in the instructions in the PR

> I'm pretty sure 

You had also said "pretty sure" before, so I thought you hadn't narrowed that down yet